fix(util): detect cycles during component dependency resolution
also added unit tests and split up the files
